requirements.txt
langchain_community
langchain_core
python-dotenv
replicate
StrOutputParser
ChatPromptTemplate